Set integration test specific resweave version (#46)
Adding a version redirect in `go.work` for `v0.0.0-test` which will be used in integration tests.
The replace for `github.com/agilitree/resweave@v0.0.0-test` will point to the local resweave folder.
This ensures the current changes for testing in resweave are being used for integration testing without having to constantly update the test version.